backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Payments
stripe==7.10.0

# AI/ML Integration
openai==1.10.0
anthropic==0.11.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
redis==5.0.1

# Label Studio SDK
label-studio-sdk==0.0.32
